Refactored FatalErrorBoundary
