+++ src/js/shot-generator-new/Editor.js
@@ -34,7 +34,7 @@ const SceneManager = require('./SceneManager')
 const GuidesView = require('./components/GuidesView').default
 const Icon = require('./Icon')
 const Toolbar = require('./components/Toolbar').default
-const FatalErrorBoundary = require('./FatalErrorBoundary')
+const FatalErrorBoundary = require('./components/FatalErrorBoundary').default
 
 const {useExportToGltf, loadCameraModel} = require('./use-export-to-gltf')
